requirements.txt
colorama==0.4.6
faiss-cpu==1.8.0.post1
langchain==0.2.17
langchain-community==0.2.19
langchain-openai==0.1.25
langchain-ollama==0.1.3
langchain-text-splitters==0.2.4
matplotlib==3.7.5
networkx==3.1
numpy==1.24.4
Pillow==10.4.0
pdfminer.six==20250324
python-dotenv==1.0.1
imageio==2.35.1
requests==2.32.4
tavily==1.1.0
gradio==4.44.0
sqlmodel==0.0.22
psycopg[binary]==3.1.18
fastapi==0.110.0
uvicorn==0.30.1
